+++ flourish/__init__.py
@@ -3,7 +3,7 @@
 from collections import defaultdict
 from datetime import date, datetime
 from importlib.machinery import SourceFileLoader
-from operator import attrgetter, itemgetter
+from operator import itemgetter
 import os
 from shutil import copyfile, rmtree
 import sys
@@ -19,6 +19,7 @@ from .source import (
     SourceFile,
     TomlSourceFile,
 )
+from .sourcelist import SourceList
 from .url import URL
 from .version import __version__    # noqa: F401
 
@@ -26,40 +27,23 @@ sys.dont_write_bytecode = True
 
 
 class Flourish(object):
-    ARGS = [
-        'source_dir',
-        'templates_dir',
-        'output_dir',
-        'sass_dir',
-        'jinja',
-    ]
-    DATA = [
-        '_assets',
-        '_cache',
-        '_source_files',
-        '_source_url',
-        '_urls',
-    ]
-    ACTIONS = [
-        '_filters',
-        '_order_by',
-        '_slice',
-    ]
-
     def __init__(
         self,
         source_dir='source',
         templates_dir='templates',
         output_dir='output',
         sass_dir='sass',
-        global_context=None,
-        **kwargs
     ):
         self.source_dir = source_dir
         self.templates_dir = templates_dir
         self.output_dir = output_dir
         self.sass_dir = sass_dir
-        self.global_context = global_context
+        self._assets = {}
+        self._cache = {}
+        self._source_files = []
+        self._source_url = None
+        self._urls = {}
+
         self.jinja = Environment(
             loader=FileSystemLoader(self.templates_dir),
             keep_trailing_newline=True,
@@ -68,28 +52,12 @@ class Flourish(object):
         self.jinja.globals['url'] = self.template_resolve_url
         self.jinja.globals['lookup'] = self.template_get
 
-        self._assets = {}
-        self._cache = {}
-        self._filters = []
-        self._order_by = []
-        self._slice = None
-        self._source_files = []
-        self._source_url = None
-        self._urls = {}
-
         if not os.path.isdir(self.source_dir):
             raise AttributeError(
                 'source_dir "%s" must exist' % self.source_dir)
 
-        for _opt in self.DATA + self.ACTIONS:
-            if _opt in kwargs:
-                if kwargs[_opt] is not None:
-                    setattr(self, _opt, kwargs[_opt])
-
         self.site_config = self._read_site_config()
-
-        if '_source_files' not in kwargs:
-            self._rescan_sources()
+        self._rescan_sources()
 
         generate = SourceFileLoader(
                 'generate', '%s/generate.py' % self.source_dir
@@ -128,18 +96,7 @@ class Flourish(object):
 
     @property
     def sources(self):
-        return self
-
-    def all(self):
-        """ Get all source documents. """
-        return self
-
-    def count(self):
-        """ Get the count of all source documents. """
-        _sources = 0
-        for source in self.sources.all():
-            _sources += 1
-        return _sources
+        return SourceList(self._source_files)
 
     def get(self, slug):
         """ Get a single source document by slug. """
@@ -154,27 +111,6 @@ class Flourish(object):
         except SourceFile.DoesNotExist:
             warnings.warn('Cannot get "%s" as it does not exist' % slug)
             return None
-
-    def filter(self, **kwargs):
-        _clone = self.clone()
-        for _key, _value in kwargs.items():
-            _clone._filters.append((_key, _value))
-        return _clone
-
-    def exclude(self, **kwargs):
-        _clone = self.clone()
-        for _key, _value in kwargs.items():
-            if '__' in _key:
-                _field, _operator = _key.split('__', 2)
-            else:
-                _field, _operator = _key, 'eq'
-            _new_operator = INVERSE_OPERATORS.get(_operator)
-            _key = '%s__%s' % (_field, _new_operator)
-            _clone._filters.append((_key, _value))
-        return _clone
-
-    def order_by(self, *args):
-        return self.clone(_order_by=args)
 
     def canonical_source_url(self, url, generator):
         # FIXME hmmmm
@@ -227,14 +163,14 @@ class Flourish(object):
 
     def generate_url(self, name, report=False):
         url = self._urls[name]
-        url.generator(self.clone(), url, self.global_context, report=report)
+        url.generator(self, url, self.global_context, report=report)
 
     def generate_path(self, path, report=False):
         handlers = self.get_handler_for_path(path)
         for key, args in handlers:
             url = self._urls[key]
             url.generator(
-                self.clone(),
+                self,
                 url,
                 self.global_context,
                 report=report,
@@ -258,24 +194,6 @@ class Flourish(object):
             copyfile(_source, _destination)
             if report:
                 print('->', _destination)
-
-    def clone(self, **kwargs):
-        for _option in self.ARGS + self.DATA + self.ACTIONS:
-            _value = getattr(self, _option)
-            if type(_value) == list:
-                kwargs.setdefault(_option, _value[:])
-            else:
-                kwargs.setdefault(_option, _value)
-        return type(self)(**kwargs)
-
-    def copy(self, **kwargs):
-        for _option in self.ARGS + self.DATA:
-            _value = getattr(self, _option)
-            if type(_value) == list:
-                kwargs.setdefault(_option, _value[:])
-            else:
-                kwargs.setdefault(_option, _value)
-        return type(self)(**kwargs)
 
     def _read_site_config(self):
         _config_file = '%s/_site.toml' % self.source_dir
@@ -327,7 +245,7 @@ class Flourish(object):
             return defaultdict(recursively_default_dict)
 
         _captured = recursively_default_dict()
-        for _source in self.all():
+        for _source in self.sources.all():
             try:
                 _date = getattr(_source, 'published')
                 _captured[_date.year][_date.month][_date.day] = 1
@@ -389,172 +307,5 @@ class Flourish(object):
         # the sort is only so the tests can compare results easily
         return sorted(_filters, key=itemgetter(_first_token))
 
-    def __len__(self):
-        return self.count()
-
-    def __getitem__(self, item):
-        if isinstance(item, slice):
-            if (item.start < 0) or (item.stop < 0):
-                raise ValueError('Cannot use negative indexes with Flourish')
-            return self.clone(_slice=item)
-        if item < 0:
-            raise ValueError('Cannot use negative indexes with Flourish')
-        _iterator = iter(self)
-        try:
-            for i in range(0, item+1):
-                obj = next(_iterator)
-        except StopIteration:
-            raise IndexError(item)
-        return obj
-        return None
-
-    def _get_filtered_sources(self):
-        _sources = []
-        for _source in self._source_files:
-            _add = True
-            if len(self._filters):
-                for _filter in self._filters:
-                    if '__' in _filter[0]:
-                        _field, _operator = _filter[0].split('__', 2)
-                    else:
-                        _field, _operator = _filter[0], 'eq'
-                    _operation = OPERATORS.get(_operator)
-                    if _operation is None:
-                        # FIXME write test and raise more useful error
-                        raise RuntimeError
-                    _is_date_filter = (
-                        _field in ['year', 'month', 'day'] and
-                        'published' in _source and
-                        type(_source['published'] == datetime)
-                    )
-
-                    try:
-                        if _is_date_filter:
-                            _test = '%02d' % getattr(
-                                _source['published'], _field)
-                        else:
-                            _test = getattr(_source, _field)
-                    except AttributeError:
-                        _test = None
-
-                    _result = _operation(_test, _filter[1])
-                    if not _result:
-                        _add = False
-            if _add:
-                _sources.append(_source)
-        if self._slice is not None:
-            _sources = _sources.__getitem__(self._slice)
-        return _sources
-
-    def __iter__(self):
-        _sources = self._get_filtered_sources()
-        for _order in self._order_by:
-            if _order[0] == '-':
-                _rev = True
-                _attr = _order[1:]
-            else:
-                _rev = False
-                _attr = _order
-            try:
-                _sources = sorted(
-                    _sources, key=attrgetter(_attr), reverse=_rev)
-            except AttributeError:
-                warnings.warn(
-                    'sorting sources by "%s" failed: '
-                    'not all sources have that attribute' % _order
-                )
-
-        return iter(_sources)
-
     def __repr__(self):
         return '<flourish.Flourish object (source=%s)>' % self.source_dir
-
-
-def _equal_or_inside(value, test):
-    if type(value) == list:
-        return test in value
-    else:
-        return value == test
-
-
-def _not_equal_or_inside(value, test):
-    return not _equal_or_inside(value, test)
-
-
-def _less_than(value, test):
-    return value < test
-
-
-def _less_than_or_equal_to(value, test):
-    return value <= test
-
-
-def _greater_than(value, test):
-    return value > test
-
-
-def _greater_than_or_equal_to(value, test):
-    return value >= test
-
-
-def _contains(value, test):
-    try:
-        if type(value) == list:
-            for _v in value:
-                if test in _v:
-                    return True
-            return False
-        else:
-            return test in value
-    except TypeError:
-        return False
-
-
-def _excludes(value, test):
-    return not _contains(value, test)
-
-
-def _inside(value, test):
-    return value in test
-
-
-def _outside(value, test):
-    return value not in test
-
-
-def _set(value, test):
-    return value is not None
-
-
-def _unset(value, test):
-    return value is None
-
-
-OPERATORS = {
-    'eq': _equal_or_inside,
-    'neq': _not_equal_or_inside,
-    'lt': _less_than,
-    'lte': _less_than_or_equal_to,
-    'gt': _greater_than,
-    'gte': _greater_than_or_equal_to,
-    'contains': _contains,
-    'excludes': _excludes,
-    'in': _inside,
-    'notin': _outside,
-    'set': _set,
-    'unset': _unset,
-}
-INVERSE_OPERATORS = {
-    'neq': 'eq',
-    'eq': 'neq',
-    'lt': 'gte',
-    'lte': 'gt',
-    'gt': 'lte',
-    'gte': 'lt',
-    'contains': 'excludes',
-    'excludes': 'contains',
-    'in': 'notin',
-    'notin': 'in',
-    'set': 'unset',
-    'unset': 'set',
-}